Motor controller now works
The connection can now be closed topside without the whole thing
failing, and the connect and disconnect buttons enable and disable
appropriately. Changed the Connection class to not initiate the
TcpClient connection upon construction - now the Connect method must be
called, which indicates whether or not the connection was successful and
displays any errors in the listbox instead of a snarky messagebox.

TODO: -Test with multiple controllers;
-Send data response;
-Process data response;
-Switch to WPF